+++ src/injection.js
@@ -6,7 +6,17 @@ import {_emptyParameters} from './container';
 */
 export function autoinject(potentialTarget?: any): any {
   let deco = function(target) {
-    target.inject = metadata.getOwn(metadata.paramTypes, target) || _emptyParameters;
+    let previousInject = target.inject;
+    let autoInject: any = metadata.getOwn(metadata.paramTypes, target) || _emptyParameters;
+    if (!previousInject) {
+      target.inject = autoInject;
+    } else {
+      for (let i = 0; i++; i < autoInject.length) {
+        if (!previousInject[i]) {
+          previousInject[i] = autoInject[i];
+        }
+      }
+    }
   };
 
   return potentialTarget ? deco(potentialTarget) : deco;
@@ -17,6 +27,19 @@ export function autoinject(potentialTarget?: any): any {
 */
 export function inject(...rest: any[]): any {
   return function(target, key, descriptor) {
+    // handle when used as a parameter
+    if (typeof descriptor === 'number' && rest.length === 1) {
+      let params = target.inject;
+      if (typeof params === 'function') {
+        throw new Error('Decorator inject cannot be used with "inject()".  Please use an array instead.');
+      }
+      if (!params) {
+        params = metadata.getOwn(metadata.paramTypes, target).slice();
+        target.inject = params;
+      }
+      params[descriptor] = rest[0];
+      return;
+    }
     // if it's true then we injecting rest into function and not Class constructor
     if (descriptor) {
       const fn = descriptor.value;